Update styleroot to use newly packaged SBP layout
The current workaround for getting the SBP layout was that Meike had to
use specially checked out branch of the suse-xsl repo. Absolutely minor
tradeoff was that only she could build SBPs properly.

We now have a package for these stylesheets, so the situation gets at
least marginally better, in that Docserv2 will be able to build this
stuff as well.