Fixed actual publication date for all documents
Previously, for xml files the automatic date stamp was used. In
consequence, instead of showing the actual publication date, with each
new build, the latest build date was shown in the document. For
adoc files we had no option to add the real publication date, and the
build date was also set as publication date. This is fixed now. Adjusted
all necessary files, DC and xml, to reflect the real publication dates
of the documents.