Fixed format issues and typos
The author used an uncommon tagging method for titles and subtitles,
using underscores and swung dashes as tags or marks. The number of these
marks needs to exactly match the number of letters of the title or
subtitle, or the entire document is messed up. Fixed that. Fixed other
format issues for consistency, and fixed typos.