Outcommented duplicated section
Section 3.5.3 Configuring Data Disks to Hold Repositories and DB exists
in big part twice in the original source, right before the section
mentioned above, there is a level 4 section with the same title but
contains only the first part of the content of the actual section 3.5.3.
I have outcommented the duplicated and uncomplete level 4 section.
Hopefully this was correct.