Copied updated adoc from author to push into main
To resolve the branch issue in GitHub, I copied the adoc from Stephen which was updated manually by Stephen with
my former edits and his content updates into my local main repo to push
it now into the official SPB repo. This is to ensure that all content
updates are there.