Date changed, slight edit of abstract
Changed the date to the actual date to have that also correct in the
docinfo file.
Edited abstract: removed the words "cloud platform" to avoid duplication
regarding the Google Cloud Platform GCP. As we have one common MAIN file
for three different cloud provider outputs, having a sentence that reads
"for the Google Cloud Platform Cloud Platform" is simply horrible. This
change however does also change the abstracts of the outputs for Azure
and AWS to read "for Azure" and "for AWS" without the addition "cloud
platform".