Several smaller edits, including the cloud variable
Changed the cloud variable for GCP to Google Cloud Platform because this
is the official name by Google. This is also to avoid the duplication
"GCP Cloud Platform" which without abbreviation would mean "Google Cloud Platform Cloud
Platform".
Removed "Cloud Platform" from the title, also to avoid duplication. And
for AWS and Azure, we don't need "Cloud Platform" explicitely. It is not
part of the proper name, and everyone working with AWS and Azure know
that these are cloud platforms.
Edited Peter's job title to Senior.
Fixed punctuation mistake.