Final edits according to Style and Spell Checker
ATTENTION: In the ifdev for GCP, I had to change the variable for cloud
from "GCP" to "Google Cloud Platform" to ensure the proper and correct title and
naming in the abstract. Of course, this needed to be changed throughout
the document.

Implemented enhancements proposed by Style Checker and Spell Checker.
Removed some filling words. Fixed grammar mistakes, typos and punctuation issues.