Implemented enhancements from Style Checker
Edited the doc following our Style Guide and implementing the
enhancements proposed by Style Checker.
Fixed more punctuation issues and smaller typos.
Fixed typo throughout document: Region Server is treated as a proper name and both
components of the name should be capitalized, with RMT server, the
server component is not part of the proper name and should not be
capitalized.