Updated SBP legal notice
We have updated the SBP legal notice section as agreed with Legal/Ciaran
Farell for better readability. Also, we created an sbp-legal-notice file
as an include to ensure that not every single document has to be touched
in case we need to change some content of the legal notice in general.