DC files: Update to new STYLEROOT/FALLBACK_STYLEROOT across all DCs
This is a bit of a temporary hack, so we can finally use the suse2021
styles at least for the HTML, but stick to existing suse2013-sbp styles
for the PDFs (for the moment).

The HTML styles have no further customizations, so taking the newer
version should be fine. The FO styles still have some unported SBP-
related features and we'll fix them when those features are ported.