Changed section titles where necessary
Since a while, we do not capitalize section titles anymore, but use
sentence-style titles. In addition, sections that describe a task or a
procedure should have a descriptive title, using the Gerund to indicate
an action or a task. Fixed that.

Also, fixed some punctuation issue and typos.

More edits/enhancements to come with style check.