Renamed files for better SEO
To improve search engine resulty, no underscores are used in file names anymore.Also for better SEO, capital letters should be avoided in file names. They are mainly used to indicate the file type DC or MAIN, and to indicate the document taype SBP or TRD, and for certain product names.