Smaller fixes and enhancements
Fixed the ID of the article as it used the ID of another already existing
SBP.
Changed title to capital letters (titles are capitalized, section
headlines are not).
Implemented smaller fixes from Style Checker: we are not using abbreviations,
such as e.g., or acronyms for product names, such as SLE. We also avoid
to use words such as once, in order to, and some others.
Added second author, changed date to June 4 because the paper will be
online on that date.