Fixing metadata for SPBs
Mainly fixed SAP version number. As we need to be able to select
main versions of SAP in the doc portal via pop-up window, we need to
limit the info in the metatag to the pure major version, without service
pack. Fixed that.
Outcommented doc type as not needed here.