Outcommented date from DC file
Adoc documents had the date in the DC file, added as an Argument for
AsciiDoc. This is not necessary anymore, as the document date will be
generated from the revhistory tag in the docinfo.xml file. To ensure
that the different date tags do not conflict, I have outcommented the
date argument from all adoc DC files.