Correct the stated reason why 538.imagick_r is slower with PGO
After the GCC 14 SBP got published, our colleague Jan Hubička
completed his detailed analysis why 538.imagick_r is slower when built
with PGO and the reason is actually different from what we stated in
the first version of the paper.  This commit replaces the incorrect
one with the correct one.